share options between composed app and server
