+++ generators/app/index.js
@@ -1,8 +1,14 @@
 const path = require("path");
 
 const EngageGenerator = require("../../lib/EngageGenerator");
+const initServerOptions = require("../server/initServerOptions");
 
 module.exports = class AppGenerator extends EngageGenerator {
+  constructor(args, options) {
+    super(args, options);
+    initServerOptions(this);
+  }
+
   initializing() {
     this.composeWith(require.resolve("../client"), {
       ...this.options,
