+++ generators/server/constants.js
@@ -1,0 +1,9 @@
+const supportedViewEngines = ["handlebars", "none"];
+const supportedTestFrameworks = ["jest"];
+const supportedDbClients = ["objection", "pg"];
+
+module.exports = {
+  supportedViewEngines,
+  supportedTestFrameworks,
+  supportedDbClients,
+};
